==> SEA_AD/step7_check_patient_subset.Rmd <==
---
title: "Check the subset of patients identified by Astro and Micro-PVM expression"
author: "Wei Sun"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: journal
    highlight: tango
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: false
    number_sections: false
  df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(width = 80)
```


# Setup

## Load R libraries
```{r load_libraries, message=FALSE, warning=FALSE}
library(data.table)
library(ggplot2)
library(ggpubr)
library(tidyr)
library(stringr)
library(reticulate)
library(zellkonverter)
library(SingleCellExperiment)
library(RColorBrewer)
library(readxl)
library(ggcorrplot)

theme_set(theme_classic())

np = import("numpy")
```

# Examine results

## Read in individual level scores

```{r fig.width=4.5, fig.height=3}

micro = np$load("output/score_ind_Micro-PVM_2.0.npy", allow_pickle = TRUE)
astro = np$load("output/score_ind_Astro_2.0.npy")

dim(astro)
astro[1:2,]

micro[,1][1:5]
micro[,3][1:5]

micro = apply(micro, 2, unlist)
dim(micro)
micro[1:2,]

table(micro[,1], astro[,1])

df = data.frame(dementia = as.factor(micro[,1]), 
                Donor.ID = as.character(micro[,3]), 
                micro = round(as.numeric(micro[,2]),7), 
                astro = as.numeric(astro[,2]))
dim(df)
df[1:2,]

ggplot(data = df, aes(x = astro, y = micro, color = dementia)) +
  geom_point() + geom_hline(yintercept=0.7) + geom_vline(xintercept=0.8)

df$subset = df$micro > 0.7 & df$astro > 0.8
table(df$subset)
```


## Read in donor infomration

```{r fig.width=3.2, fig.height=2.4}
donor = fread("data/sea-ad_cohort_donor_metadata_simplified.tsv")
dim(donor)
donor[1:2,]

table(donor$Alzheimers)
table(df$Donor.ID %in% donor$`Donor ID`)

donor = donor[match(df$Donor.ID, donor$`Donor ID`),]
dim(donor)
donor[1:2,]

df = cbind(df, donor)
dim(df)
df[1:5,]

table(df$`Donor ID` == df$`Donor ID`)

table(df$dementia, df$subset)

df_dm = df[df$dementia == 1, ]
table(df_dm$subset, df_dm$Sex)

table(df_dm$subset, df_dm$`APOE Genotype`)
chisq.test(df_dm$subset, df_dm$`APOE Genotype`)

table(df_dm$subset, df_dm$`Primary Study Name`)
chisq.test(df_dm$subset, df_dm$`Primary Study Name`)

tapply(df_dm$`Years of education`, df_dm$subset, summary)
wilcox.test(df_dm$`Years of education` ~ df_dm$subset)

table(df_dm$subset, df_dm$Thal)
chisq.test(df_dm$subset, df_dm$Thal)

table(df_dm$subset, df_dm$Braak)
chisq.test(df_dm$subset, df_dm$Braak)

df_dm$Braak_group = rep("II-IV", nrow(df_dm))
df_dm$Braak_group[df_dm$Braak == "Braak V"] = "V"
df_dm$Braak_group[df_dm$Braak == "Braak VI"] = "VI"

tb1 = table(df_dm$subset, df_dm$Braak_group)
chisq.test(df_dm$subset, df_dm$Braak_group)

df1 = as.data.frame(tb1)
names(df1)[1:2] = c("Donor subset", "Braak group")
df1

df1$group = rep("Dementia-s", nrow(df1))
df1$group[df1$`Donor subset` == "TRUE"] = "Dementia-i"
df1$group = factor(df1$group, levels = c("Dementia-s", "Dementia-i"))
df1

ggplot(data=df1, aes(x = `Braak group`, y = Freq, fill=group)) +
  geom_bar(stat = "identity", position = position_dodge()) + 
  scale_fill_manual(values =c("#FF8C00EE", "purple"))
```

## Read in neuropathological features

```{r fig.width=9, fig.height=9}
features = read_excel("data/media-2.xlsx")
dim(features)
features[1:2,1:4]

grey_matter = grep("Grey matter", names(features))
length(grey_matter)
grey_matter[1:4]
names(features)[grey_matter[1:4]]

features = features[,-grey_matter]
dim(features)
features[1:2,1:4]

gAT8 = setdiff(grep("AT8", names(features)), 
               grep("pTDP43", names(features)))

names(features)[gAT8]

cor_AT8 = cor(features[,gAT8], method = "spearman")
summary(c(cor_AT8[upper.tri(cor_AT8)]))

ggcorrplot(cor_AT8) + theme(axis.text.x = element_blank()) + 
  scale_fill_gradient2(low = "blue", mid = "white", high = "red", 
                       midpoint = 0.5, limits = c(-0.1, 1))
```

```{r fig.width=10, fig.height=10}

g6e10 = setdiff(grep("6e10", names(features)), 
                grep("Iba1", names(features)))

names(features)[g6e10]

cor_6e10 = cor(features[,g6e10], method = "spearman")
summary(c(cor_6e10[upper.tri(cor_6e10)]))

ggcorrplot(cor_6e10) + theme(axis.text.x = element_blank()) + 
  scale_fill_gradient2(low = "blue", mid = "white", high = "red", 
                       midpoint = 0.7, limits = c(0.3, 1))
```

```{r fig.width=12, fig.height=12}

gI6 = intersect(grep("6e10", names(features)), 
                grep("Iba1", names(features)))

names(features)[gI6]

cor_I6 = cor(features[,gI6], method = "spearman")
summary(c(cor_I6[upper.tri(cor_I6)]))

ggcorrplot(cor_I6) + theme(axis.text.x = element_blank()) + 
  scale_fill_gradient2(low = "blue", mid = "white", high = "red", 
                       midpoint = 0.5, limits = c(-0.1, 1))

gI6 = grep("percent of Iba1 and 6e10 positive", names(features))
cor_I6 = cor(features[,gI6], method = "spearman")
summary(c(cor_I6[upper.tri(cor_I6)]))
```

Most features are strongly correlated. So we just use the "percent of xxx positive" in the following analysis

```{r fig.width=8, fig.height=8}

gAT8 = grep("percent AT8 positive", names(features))
names(features)[gAT8]
cor_AT8 = cor(features[,gAT8], method = "spearman")
summary(c(cor_AT8[upper.tri(cor_AT8)]))

g6e10 = grep("percent 6e10 positive", names(features))
names(features)[g6e10]
cor_6e10 = cor(features[,g6e10], method = "spearman")
summary(c(cor_6e10[upper.tri(cor_6e10)]))

gI6 = grep("percent of Iba1 and 6e10 positive", names(features))
names(features)[gI6]
cor_I6 = cor(features[,gI6], method = "spearman")
summary(c(cor_I6[upper.tri(cor_I6)]))

feature_df = as.data.frame(features[,c(1, gAT8, g6e10, gI6)])
dim(feature_df)
feature_df[1:2,1:3]
names(feature_df)
feature_df$pTau = rowMeans(features[,gAT8])
feature_df$amyloid_beta = rowMeans(features[,g6e10])
feature_df$Iba1_amyloid_beta = rowMeans(features[,gI6])

names(feature_df) = gsub("percent of ", "", names(feature_df), fixed=TRUE)
names(feature_df) = gsub("percent ", "", names(feature_df))
names(feature_df) = gsub("co-localized objects_", "", names(feature_df), 
                         fixed=TRUE)
names(feature_df) = gsub("area_", "", names(feature_df), fixed=TRUE)

cor_mat = cor(feature_df[,-1], method = "spearman")
summary(c(cor_mat[upper.tri(cor_mat)]))
ggcorrplot(cor_mat) + theme(axis.text.x = element_blank()) + 
  scale_fill_gradient2(low = "blue", mid = "white", high = "red", 
                       midpoint = 0.75, limits = c(0.5, 1))

table(df$Donor.ID %in% feature_df$`Donor ID`)
df_test = merge(df, feature_df)
dim(df_test)

table(df_test$dementia, df_test$subset)
df_test$group = rep("No dementia", nrow(df_test))
df_test$group[df_test$dementia == 1] = "Dementia-s"
df_test$group[which(df_test$subset)] = "Dementia-i"
table(df_test$group)
```


```{r fig.width=4, fig.height=2.6}
df_test$group = factor(df_test$group, 
                       levels = c("No dementia", "Dementia-s", 
                                  "Dementia-i"))

compare_list = list(c("No dementia", "Dementia-s"), 
                    c("No dementia", "Dementia-i"), 
                    c("Dementia-s", "Dementia-i"))

g1 = ggplot(df_test, aes(x=group, y=pTau, color = group)) + 
  geom_boxplot(outlier.shape=-1) + 
  scale_color_manual(values = c("deepskyblue4", "chocolate2", "purple")) + 
  geom_jitter(width = 0.2, size = 1.5, alpha = 0.9, height=0) + 
  stat_compare_means(comparisons = compare_list, method = "wilcox.test") + 
  labs(x = NULL, y = "pTau") +
  theme(axis.text.x = element_blank(), legend.text = element_text(size = 11))  

g2 = ggplot(df_test, aes(x=group, y=amyloid_beta, color = group)) + 
  geom_boxplot(outlier.shape=-1) + 
  scale_color_manual(values = c("deepskyblue4", "chocolate2", "purple")) + 
  geom_jitter(width = 0.2, size = 1.5, alpha = 0.9, height=0) + 
  stat_compare_means(comparisons = compare_list, method = "wilcox.test") + 
  labs(x = NULL, y = "Amyloid beta") +
  theme(axis.text.x = element_blank(), legend.text = element_text(size = 11))  

g3 = ggplot(df_test, aes(x=group, y=Iba1_amyloid_beta, color = group)) + 
  geom_boxplot(outlier.shape=-1) + 
  scale_color_manual(values = c("deepskyblue4", "chocolate2", "purple")) + 
  geom_jitter(width = 0.2, size = 1.5, alpha = 0.9, height=0) + 
  stat_compare_means(comparisons = compare_list, method = "wilcox.test") + 
  labs(x = NULL, y = "Iba1 and Amyloid beta") +
  theme(axis.text.x = element_blank(), legend.text = element_text(size = 11))  
g3
```

```{r fig.width=5, fig.height=2.9}
ggarrange(g1, g2, nrow = 1, common.legend = TRUE)
```

```{r fig.width=2.5, fig.height=2.3}
g1 = ggplot(df_test, aes(x=pTau, y=amyloid_beta, color = group)) + 
  geom_point(size=2, alpha=0.9) + 
  scale_color_manual(values = c("deepskyblue4", "chocolate2", "purple")) + 
  labs(y = "Amyloid beta") + 
  theme(legend.text = element_text(size = 11),
        legend.position = "none")

g2 = ggplot(df_test, aes(x=pTau, y=Iba1_amyloid_beta, color = group)) + 
  geom_point(size=2, alpha=0.9) + 
  scale_color_manual(values = c("deepskyblue4", "chocolate2", "purple")) + 
  labs(y = "Iba1 and Amyloid beta") + 
  theme(legend.text = element_text(size = 11), 
        legend.position = "none")

g3 = ggplot(df_test, aes(x=amyloid_beta, y=Iba1_amyloid_beta, 
                         color = group)) + 
  geom_point(size=2, alpha=0.9) + 
  scale_color_manual(values = c("deepskyblue4", "chocolate2", "purple")) + 
  labs(x = "Amyloid beta", y = "Iba1 and Amyloid beta") + 
  theme(legend.text = element_text(size = 11),
        legend.position = "none")

g1
g2
g3
```

# Session information

```{r}
gc()

sessionInfo()
```
